chore: shared poll implementation
